requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anomalib==2.1.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
apturl==0.5.2
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
autogen-agentchat==0.4.2
autogen-core==0.4.2
autogen-ext==0.4.2
Automat==20.2.0
av==13.1.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blessings==1.7
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==25.1.4
cobble==0.1.4
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.0
cryptography==44.0.0
cupshelpers==1.0
cycler==0.12.1
datasets==3.0.1
dbus-python==1.2.18
debugpy==1.8.12
decorator==5.1.1
decord==0.6.0
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docstring_parser==0.17.0
duplicity==0.8.21
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.0
fasteners==0.14.1
ffmpy==0.4.0
filelock==3.14.0
fonttools==4.54.1
FrEIA==0.2
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.2
fvcore==0.1.5.post20221221
gpustat==0.6.0
gradio==4.44.0
gradio_client==1.3.0
greenlet==3.1.1
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.25.1
hyperlink==21.0.0
idna==3.3
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==21.3.0
iniconfig==2.1.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.8.2
joblib==1.4.2
jsonargparse==4.40.1
jsonpatch==1.32
jsonpointer==2.0
jsonref==1.1.0
jsonschema==3.2.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.15.0
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==2.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mutagen==1.47.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
openai==1.59.9
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.29.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
paramiko==2.9.3
parso==0.8.4
pathvalidate==3.2.3
pdfminer.six==20240706
Pebble==5.1.1
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.2.2
playwright==1.49.1
plotly==5.23.0
pluggy==1.6.0
portalocker==2.10.1
prompt_toolkit==3.0.50
protobuf==4.25.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.20.0
pycups==2.0.1
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==8.4.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
python-multipart==0.0.10
python-pptx==1.0.2
pytorch-lamb==1.0.0
pytorch-lightning==2.5.2
pytorchvideo==0.1.5
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.1
regex==2024.9.11
reportlab==3.6.8
requests==2.32.3
rich==13.8.1
rich-argparse==1.7.1
ruff==0.6.7
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sos==4.8.2
soupsieve==2.6
SpeechRecognition==3.14.0
speedtest-cli==2.1.3
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.2
systemd-python==234
tabulate==0.9.0
tenacity==8.5.0
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.5.10
tiktoken==0.8.0
timm==1.0.19
tokenizers==0.20.0
tomli==2.2.1
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.8.0
torchvision==0.19.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.0
trash-cli==0.17.1.14
triton==3.0.0
Twisted==22.1.0
typer==0.12.5
typeshed_client==2.8.2
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
uv==0.8.4
uvicorn==0.30.6
virtualenv==20.26.2
wadllib==1.3.6
wcwidth==0.2.13
websockets==12.0
wrapt==1.17.2
xdg==5
xkit==0.0.0
XlsxWriter==3.2.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.13.0
youtube-transcript-api==0.6.3
yt-dlp==2024.8.6
zipp==3.21.0
zope.interface==5.4.0
